# docs/swagger.yaml
definitions:
...
  dao.App:
    properties:
      app_id:
        type: string
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      update_at:
        type: string
      white_ips:
        type: string
    type: object
...
  dto.APPAddHttpInput:
    properties:
      app_id:
        type: string
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      white_ips:
        type: string
    required:
    - app_id
    - name
    type: object
  dto.APPListItemOutput:
    properties:
      app_id:
        type: string
      create_at:
        type: string
      id:
        type: integer
      is_delete:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      real_qpd:
        type: integer
      real_qps:
        type: integer
      secret:
        type: string
      update_at:
        type: string
      white_ips:
        type: string
    type: object
  dto.APPListOutput:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.APPListItemOutput'
        type: array
      total:
        type: integer
    type: object
  dto.APPUpdateHttpInput:
    properties:
      app_id:
        type: string
      id:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      white_ips:
        type: string
    required:
    - id
    - name
    - secret
    type: object
...
  dto.StatisticsOutput:
    properties:
      today:
        items:
          type: integer
        type: array
      yesterday:
        items:
          type: integer
        type: array
    required:
    - today
    - yesterday
    type: object
  middleware.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      errmsg:
        type: string
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
paths:
...
  /app/app_add:
    post:
      consumes:
      - application/json
      description: 租户添加
      operationId: /app/app_add
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.APPAddHttpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 租户添加
      tags:
      - 租户管理
  /app/app_delete:
    get:
      consumes:
      - application/json
      description: 租户删除
      operationId: /app/app_delete
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 租户删除
      tags:
      - 租户管理
  /app/app_detail:
    get:
      consumes:
      - application/json
      description: 租户详情
      operationId: /app/app_detail
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dao.App'
              type: object
      summary: 租户详情
      tags:
      - 租户管理
  /app/app_list:
    get:
      consumes:
      - application/json
      description: 租户列表
      operationId: /app/app_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页多少条
        in: query
        name: page_size
        required: true
        type: string
      - description: 页码
        in: query
        name: page_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.APPListOutput'
              type: object
      summary: 租户列表
      tags:
      - 租户管理
  /app/app_stat:
    get:
      consumes:
      - application/json
      description: 租户统计
      operationId: /app/app_stat
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StatisticsOutput'
              type: object
      summary: 租户统计
      tags:
      - 租户管理
  /app/app_update:
    post:
      consumes:
      - application/json
      description: 租户更新
      operationId: /app/app_update
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.APPUpdateHttpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 租户更新
      tags:
      - 租户管理
...
swagger: "2.0"